pizza_shop/pyproject.toml
[tool.poetry]
name = "pizza-shop"
version = "0.1.0"
description = ""
authors = ["Andre Bossard <anbossar@microsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
reactpy = "^1.0.2"
opentelemetry-exporter-zipkin = "^1.21.0"
uvicorn = "^0.24.0.post1"
opentelemetry-instrumentation-fastapi = "^0.42b0"
pizzalibrary = {path = "../pizzalibrary", develop = true}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
